requirements.txt
absl-py==2.1.0
astunparse==1.6.3
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imutils==0.5.4
joblib==1.4.2
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mutagen==1.47.0
namex==0.0.8
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pip==24.0
progressbar==2.5
protobuf==4.25.3
psutil==5.9.8
pycryptodomex==3.20.0
pygame==2.5.2
Pygments==2.18.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yt-dlp==2024.5.27
